+++ src/ui_components/ConfirmDialog.tsx
@@ -1,0 +1,49 @@
+import * as React from 'react';
+import Button from '@mui/material/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
+
+export interface ConfirmDialogProps {
+    open: boolean;
+    title: string;
+    text: string;
+    onConfirm: () => void;
+    onCancel: () => void;
+}
+
+export default function ConfirmDialog(props: ConfirmDialogProps) {
+
+  return (
+      <>
+      <Dialog
+        open={props.open}
+        onClose={props.onCancel}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+      >
+        <DialogTitle id="alert-dialog-title">
+          {props.title}
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText id="alert-dialog-description">
+          {props.text}
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={props.onCancel}>Cancel</Button>
+          <Button onClick={props.onConfirm} autoFocus>
+            Confirm
+          </Button>
+        </DialogActions>
+      </Dialog>
+    </>
+  );
+}
+
+ConfirmDialog.defaultProps = {
+    title: 'Confirm',
+    text: 'Are you sure?',
+};