+++ src/ui_components/LayerList/util/LayerPropertyForm.tsx
@@ -6,7 +6,7 @@ import {
 import React, { useRef } from 'react';
 import ColorPicker from './input/ColorPicker';
 import { Box, ListItem, Paper, Slider, TextField, Typography } from '@mui/material';
-import {useCallback} from 'react';
+import { useCallback } from 'react';
 
 export type paintPropsType =
 	| CircleLayerSpecification['paint']
@@ -61,67 +61,76 @@ export default function LayerPropertyForm({ paintProps = {}, setPaintProps }: Pr
 	const key = useRef(Math.round(Math.random() * 10000000000));
 	//const onChange = (event) => {};
 
-	const getFormInputByType = useCallback((key: string) => {
-		if (mapPropKeyToFormInputTypeKeys.indexOf(key) !== -1) {
-			const label = (
-				<Typography id={key + '_label'} gutterBottom>
-					{key}
-				</Typography>
-			);
-			switch (mapPropKeyToFormInputType[key]) {
-				case 'slider':
-					return (
-						<>
-							{label}
-							<Slider
-								{...inputPropsByPropKey[key]}
-								inputProps={{ inputMode: 'decimal', pattern: '[0-9]*' }}
-								value={paintProps[key]}
-								valueLabelDisplay="auto"
-								onChange={(_ev: Event, value: number) => {
-									if (value) {
-										setPaintProps((current) => ({ ...current, [key]: value }));
-									}
-								}}
-							/>
-						</>
-					);
-					break;
-				case 'numberfield':
-					return (
-						<>
-							{label}
-							<TextField
-								inputProps={{ inputMode: 'decimal', pattern: '[0-9]*' }}
-								value={paintProps[key]}
-								onChange={(ev: React.ChangeEvent<HTMLInputElement>) => {
-									if (ev?.target?.value) {
-										setPaintProps((current) => ({ ...current, [key]: parseInt(ev.target.value) }));
-									}
-								}}
-							/>
-						</>
-					);
-					break;
-				case 'colorpicker':
-					return (
-						<>
-							{label}
-							<Box sx={{ '& > div': { width: 'initial !important' } }}>
-								<ColorPicker
-									key={key}
+	const getFormInputByType = useCallback(
+		(key: string) => {
+			if (
+				mapPropKeyToFormInputTypeKeys.indexOf(key) !== -1 &&
+				(typeof paintProps[key] === 'number' || typeof paintProps[key] === 'string')
+			) {
+				const label = (
+					<Typography id={key + '_label'} gutterBottom>
+						{key}
+					</Typography>
+				);
+				switch (mapPropKeyToFormInputType[key]) {
+					case 'slider':
+						return (
+							<React.Fragment key={key}>
+								{label}
+								<Slider
+									{...inputPropsByPropKey[key]}
+									inputProps={{ inputMode: 'decimal', pattern: '[0-9]*' }}
 									value={paintProps[key]}
-									propKey={key}
-									setPaintProps={setPaintProps}
+									valueLabelDisplay="auto"
+									onChange={(_ev: Event, value: number) => {
+										if (value) {
+											setPaintProps((current) => ({ ...current, [key]: value }));
+										}
+									}}
 								/>
-							</Box>
-						</>
-					);
-					break;
+							</React.Fragment>
+						);
+						break;
+					case 'numberfield':
+						return (
+							<React.Fragment key={key}>
+								{label}
+								<TextField
+									inputProps={{ inputMode: 'decimal', pattern: '[0-9]*' }}
+									value={paintProps[key]}
+									onChange={(ev: React.ChangeEvent<HTMLInputElement>) => {
+										if (ev?.target?.value) {
+											setPaintProps((current) => ({
+												...current,
+												[key]: parseInt(ev.target.value),
+											}));
+										}
+									}}
+								/>
+							</React.Fragment>
+						);
+						break;
+					case 'colorpicker':
+						return (
+							<React.Fragment key={key}>
+								{label}
+								<Box sx={{ '& > div': { width: 'initial !important' } }}>
+									<ColorPicker
+										key={key}
+										value={paintProps[key]}
+										propKey={key}
+										setPaintProps={setPaintProps}
+									/>
+								</Box>
+							</React.Fragment>
+						);
+						break;
+				}
 			}
-		}
-		return null;
-	},[paintProps]);
+			return null;
+		},
+		[paintProps]
+	);
 
 	return (
 		<>